Factor out apps and create storybooks for them.
