+++ packages/konsole-app/package.json
@@ -66,12 +66,14 @@
     "react-router": "^5.2.1",
     "react-router-dom": "^5.3.0",
     "react-virtualized": "^9.22.3",
+    "storybook-react-router": "^1.0.8",
     "string-hash": "^1.1.3",
     "superagent": "^6.1.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@dxos/eslint-plugin": "^1.0.14",
+    "@storybook/addon-actions": "^6.3.8",
     "@storybook/addon-knobs": "^6.3.0",
     "@storybook/react": "^6.3.7",
     "@types/faker": "^4.1.12",
@@ -82,6 +84,7 @@
     "@types/react-router": "^5.1.16",
     "@types/react-router-dom": "^5.1.9",
     "@types/react-virtualized": "^9.21.13",
+    "@types/storybook-react-router": "^1.0.1",
     "@types/string-hash": "^1.1.1",
     "@types/superagent": "^4.1.10",
     "@typescript-eslint/eslint-plugin": "^4.19.0",
