+++ packages/konsole-app/src/apps/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2021 DXOS.org
+//
+
+export * from './App';
+export * from './Debug';
