+++ packages/konsole-app/src/index.ts
@@ -2,6 +2,7 @@
 // Copyright 2021 DXOS.org
 //
 
+export * from './apps';
 export * from './components';
 export * from './containers';
 export * from './hooks';
