+++ packages/konsole-app/src/main.tsx
@@ -5,73 +5,14 @@
 import debug from 'debug';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {
-  HashRouter as Router,
-  Redirect,
-  Route,
-  Switch,
-  useHistory,
-  useParams
-} from 'react-router-dom';
 
-import { CssBaseline, ThemeProvider } from '@mui/material';
+import { ApiFactory } from '@dxos/registry-api';
 
-import { IRegistryApi, ApiFactory } from '@dxos/registry-api';
-
-import { Container, Sidebar } from './components';
+import { App } from './apps';
 import { loadConfig } from './config';
-import { IConfig, ConfigContext, RegistryContext } from './hooks';
+import { IConfig } from './hooks';
 import { panels } from './panels';
 import { createCustomTheme } from './theme';
-
-/**
- * Main app container.
- */
-const Main = () => {
-  const history = useHistory();
-  const { panel }: { panel: string } = useParams();
-
-  return (
-    <Container
-      sidebar={
-        <Sidebar
-          panels={panels}
-          selected={`/${panel}`}
-          onSelect={path => history.push(path)}
-        />
-      }
-    >
-      <>
-        {panels.map(({ path, component }) => (
-          <Route
-            key={path}
-            path={path}
-            component={component}
-          />
-        ))}
-      </>
-    </Container>
-  );
-};
-
-// TODO(burdon): Cache panels so that they are not rendered each time.
-const App = ({ config, registryApi }: { config: IConfig, registryApi: IRegistryApi }) => (
-  <ConfigContext.Provider value={config}>
-    <RegistryContext.Provider value={registryApi}>
-      <ThemeProvider theme={createCustomTheme(config)}>
-        <CssBaseline />
-        <Router>
-          <Switch>
-            <Route path='/:panel'>
-              <Main />
-            </Route>
-            <Redirect to={panels[0].path} />
-          </Switch>
-        </Router>
-      </ThemeProvider>
-    </RegistryContext.Provider>
-  </ConfigContext.Provider>
-);
 
 /**
  * React app bootstrap (providers and top-level routes).
@@ -81,12 +22,15 @@ const start = async (config: IConfig) => {
   const log = debug('dxos:console:main');
   log('Starting...', { config });
 
+  const theme = createCustomTheme(config);
   const registryApi = await ApiFactory.createRegistryApi(config.services.dxns.server);
 
   ReactDOM.render((
     <App
       config={config}
       registryApi={registryApi}
+      panels={panels}
+      theme={theme}
     />
   ), document.getElementById('root'));
 };
