+++ packages/konsole-app/src/types.ts
@@ -2,6 +2,9 @@
 // Copyright 2021 DXOS.org
 //
 
+
+import { ComponentType } from 'react';
+
 import { CID } from '@dxos/registry-api';
 
 export interface IRecordType {
@@ -26,4 +29,11 @@ export interface IService {
   ports?: string // TODO(burdon): Array of numbers?
   memory?: number
   cpu?: number
+}
+
+export interface IPanel {
+  path: string
+  label: string
+  component: ComponentType
+  icon?: ComponentType
 }
